requirements.txt
certifi==2018.11.29
chardet==3.0.4
cycler==0.10.0
idna==2.8
kiwisolver==1.0.1
matplotlib==3.0.2
numpy==1.16.0
pandas==0.24.0
patsy==0.5.1
postgres==2.2.2
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pyparsing==2.3.1
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
SQLAlchemy>=1.3.0
urllib3>=1.24.2
statsmodels==0.10.1
